+++ packages/stargate/src/stargateclient.ts
@@ -22,7 +22,7 @@ import {
   TxExtension,
 } from "./modules";
 import { QueryClient } from "./queryclient";
-import { SearchTxQuery } from "./search";
+import { isSearchTxQueryArray, SearchTxQuery } from "./search";
 
 export class TimeoutError extends Error {
   public readonly txId: string;
@@ -386,8 +386,14 @@ export class StargateClient {
     let rawQuery: string;
     if (typeof query === "string") {
       rawQuery = query;
-    } else if (Array.isArray(query)) {
-      rawQuery = query.map((t) => `${t.key}='${t.value}'`).join(" AND ");
+    } else if (isSearchTxQueryArray(query)) {
+      rawQuery = query
+        .map((t) => {
+          // numeric values must not have quotes https://github.com/cosmos/cosmjs/issues/1462
+          if (typeof t.value === "string") return `${t.key}='${t.value}'`;
+          else return `${t.key}=${t.value}`;
+        })
+        .join(" AND ");
     } else {
       throw new Error("Got unsupported query type. See CosmJS 0.31 CHANGELOG for API breaking changes here.");
     }